+++ packages/edition-slices/edition-puzzle-tiles.stories.js
@@ -1,0 +1,4 @@
+import { showcaseConverter } from "@times-components/storybook";
+import showcase from "./edition-puzzle-tiles.showcase";
+
+showcaseConverter(module, showcase);