+++ packages/edition-slices/edition-puzzles.stories.js
@@ -1,0 +1,4 @@
+import { showcaseConverter } from "@times-components/storybook";
+import showcase from "./edition-puzzles.showcase";
+
+showcaseConverter(module, showcase);