+++ packages/edition-slices/src/slices/puzzle/styles.js
@@ -1,27 +1,22 @@
-import { spacing, colours, fonts } from "@times-components/styleguide";
+import { editionBreakpoints, spacing } from "@times-components/styleguide";
 
-const styles = {
-  body: {
-    alignSelf: "flex-end",
-    width: "85%"
-  },
-  header: {
-    paddingLeft: spacing(3),
-    paddingRight: spacing(3),
-    paddingTop: spacing(3),
-    width: "85%"
-  },
-  headLine: {
-    fontFamily: fonts.headline,
-    fontSize: 25
-  },
-  main: {
-    backgroundColor: colours.functional.border,
+const smallBreakpointStyles = {
+  container: {
     flex: 1,
-    marginLeft: spacing(2),
-    marginRight: spacing(2),
-    marginTop: spacing(2)
+    paddingHorizontal: spacing(2)
   }
 };
 
-export default styles;
+const mediumBreakpointStyles = {
+  container: {
+    flex: 1,
+    flexDirection: "row",
+    paddingHorizontal: spacing(3),
+    paddingTop: spacing(2)
+  }
+};
+
+export default breakpoint =>
+  breakpoint === editionBreakpoints.small
+    ? smallBreakpointStyles
+    : mediumBreakpointStyles;