+++ packages/edition-slices/src/tiles/shared/index.js
@@ -3,5 +3,13 @@ import TileImage from "./tile-image";
 import TileLink from "./tile-link";
 import TileSummary from "./tile-summary";
 import withTileTracking from "./tile-tracking-events";
+import withPuzzleTileTracking from "./puzzle-tracking-events";
 
-export { getCrop, TileImage, TileLink, TileSummary, withTileTracking };
+export {
+  getCrop,
+  TileImage,
+  TileLink,
+  TileSummary,
+  withPuzzleTileTracking,
+  withTileTracking
+};