+++ packages/edition-slices/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.44.0](https://github.com/newsuk/times-components/compare/@times-components/edition-slices@0.43.0...@times-components/edition-slices@0.44.0) (2019-03-21)
+
+
+### Features
+
+* Design review updates (slices) ([#1817](https://github.com/newsuk/times-components/issues/1817)) ([459acba](https://github.com/newsuk/times-components/commit/459acba))
+
+
+
+
+
+# [0.43.0](https://github.com/newsuk/times-components/compare/@times-components/edition-slices@0.42.0...@times-components/edition-slices@0.43.0) (2019-03-20)
+
+
+### Features
+
+* Puzzles 768px breakpoint ([#1811](https://github.com/newsuk/times-components/issues/1811)) ([fe1c47a](https://github.com/newsuk/times-components/commit/fe1c47a))
+
+
+
+
+
 # [0.42.0](https://github.com/newsuk/times-components/compare/@times-components/edition-slices@0.41.0...@times-components/edition-slices@0.42.0) (2019-03-19)